menu extraido e aplicado, arquivos renomeados pra snake_case
