+++ lib/components/side_menu.dart
@@ -1,0 +1,120 @@
+import 'package:flutter/material.dart';
+import 'package:npj/pages/assisted_clients.dart';
+import 'package:npj/pages/home.dart';
+import 'package:npj/pages/reports.dart';
+import 'package:npj/pages/schedule.dart';
+import 'package:npj/pages/settings.dart';
+import 'package:npj/pages/statistics.dart';
+
+class SideMenu extends StatelessWidget {
+  const SideMenu({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Theme(
+      data: Theme.of(context).copyWith(
+        canvasColor: const Color.fromARGB(255, 24, 18, 43),
+      ),
+      child: Drawer(
+        // Add a ListView to the drawer. This ensures the user can scroll
+        // through the options in the drawer if there isn't enough vertical
+        // space to fit everything.
+        child: ListView(
+          // Important: Remove any padding from the ListView.
+          padding: EdgeInsets.zero,
+          children: [
+            Container(
+              height: 120,
+              width: 100,
+              margin: const EdgeInsets.all(25),
+              decoration: const BoxDecoration(
+                  color: Color.fromARGB(255, 24, 18, 43),
+                  image: DecorationImage(
+                      image: AssetImage("../assets/imagens/gpro.png"),
+                      fit: BoxFit.fill)),
+              child: const Text(''),
+            ),
+            ListTile(
+              leading: const Icon(Icons.home, color: Colors.white),
+              title: const Text('Menu', style: TextStyle(color: Colors.white)),
+              onTap: () {
+                // Update the state of the app
+                // ...
+                // Then close the drawer
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(builder: (context) => const HomePage()),
+                );
+              },
+            ),
+            ListTile(
+              leading: const Icon(Icons.menu, color: Colors.white),
+              title: const Text('Processos',
+                  style: TextStyle(color: Colors.white)),
+              onTap: () {
+                Navigator.pop(context);
+              },
+            ),
+            ListTile(
+              leading: const Icon(Icons.person, color: Colors.white),
+              title: const Text('Assistidos',
+                  style: TextStyle(color: Colors.white)),
+              onTap: () {
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(
+                      builder: (context) => const AssistidosPage()),
+                );
+              },
+            ),
+            ListTile(
+              leading: const Icon(Icons.download, color: Colors.white),
+              title: const Text('Relatórios',
+                  style: TextStyle(color: Colors.white)),
+              onTap: () {
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(builder: (context) => const ReportsPage()),
+                );
+              },
+            ),
+            ListTile(
+              leading: const Icon(Icons.calendar_month, color: Colors.white),
+              title:
+                  const Text('Agenda', style: TextStyle(color: Colors.white)),
+              onTap: () {
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(builder: (context) => const SchedulePage()),
+                );
+              },
+            ),
+            ListTile(
+              leading: const Icon(Icons.bar_chart, color: Colors.white),
+              title: const Text('Estatisticas',
+                  style: TextStyle(color: Colors.white)),
+              onTap: () {
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(
+                      builder: (context) => const StatisticsPage()),
+                );
+              },
+            ),
+            ListTile(
+              leading: const Icon(Icons.settings, color: Colors.white),
+              title: const Text('Configurações',
+                  style: TextStyle(color: Colors.white)),
+              onTap: () {
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(builder: (context) => const SettingsPage()),
+                );
+              },
+            )
+          ],
+        ),
+      ),
+    );
+  }
+}
