+++ lib/pages/assisted_process.dart
@@ -1,18 +1,13 @@
 import 'package:data_table_2/data_table_2.dart';
 import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:npj/pages/home.dart';
+import 'package:npj/components/side_menu.dart';
 import 'package:npj/pages/add_process.dart';
-import 'package:npj/pages/edit_processo.dart';
-import 'package:npj/pages/assistidos.dart';
-import 'package:npj/pages/reports.dart';
-import 'package:npj/pages/schedule.dart';
-import 'package:npj/pages/settings.dart';
-import 'package:npj/pages/statistics.dart';
+import 'package:npj/pages/edit_process.dart';
 import 'package:npj/services/firebase_service.dart';
 
 class AssistedProcessPage extends StatelessWidget {
-  const AssistedProcessPage({Key? key});
+  const AssistedProcessPage({super.key});
 
   static const appTitle = 'GProJuridico';
 
@@ -103,7 +98,7 @@ class _MyAssistedPageState extends State<MyAssistedPage> {
       ),
       body: Column(
         children: [
-          //Header com "ASSISTIDO"
+          //Header com "PROCESSOS"
           Container(
             height: 50,
             color: const Color.fromARGB(255, 50, 39, 85),
@@ -113,7 +108,7 @@ class _MyAssistedPageState extends State<MyAssistedPage> {
                   child: Align(
                     alignment: Alignment.center,
                     child: Text(
-                      "ASSISTIDOS",
+                      "PROCESSOS",
                       style: TextStyle(
                         fontSize: 22,
                         color: Colors.white,
@@ -294,133 +289,7 @@ class _MyAssistedPageState extends State<MyAssistedPage> {
       ),
       //Fim ListView
       //Inicio Menu Lateral
-      drawer: Theme(
-        data: Theme.of(context).copyWith(
-          canvasColor: const Color.fromARGB(255, 24, 18, 43),
-        ),
-        child: Drawer(
-          // Add a ListView to the drawer. This ensures the user can scroll
-          // through the options in the drawer if there isn't enough vertical
-          // space to fit everything.
-          child: ListView(
-            // Important: Remove any padding from the ListView.
-            padding: EdgeInsets.zero,
-            children: [
-              Container(
-                height: 120,
-                width: 100,
-                margin: const EdgeInsets.all(25),
-                decoration: const BoxDecoration(
-                    color: Color.fromARGB(255, 24, 18, 43),
-                    image: DecorationImage(
-                        image: AssetImage("../assets/imagens/gpro.png"),
-                        fit: BoxFit.fill)),
-                child: const Text(''),
-              ),
-              ListTile(
-                leading: const Icon(Icons.home, color: Colors.white),
-                title:
-                    const Text('Menu', style: TextStyle(color: Colors.white)),
-                onTap: () {
-                  // Update the state of the app
-                  // ...
-                  // Then close the drawer
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(builder: (context) => const HomePage()),
-                  );
-                },
-              ),
-              ListTile(
-                leading: const Icon(Icons.menu, color: Colors.white),
-                title: const Text('Processos',
-                    style: TextStyle(color: Colors.white)),
-                onTap: () {
-                  // Update the state of the app
-                  // ...
-                  // Then close the drawer
-                  Navigator.pop(context);
-                },
-              ),
-              ListTile(
-                leading: const Icon(Icons.person, color: Colors.white),
-                title: const Text('Assistidos',
-                    style: TextStyle(color: Colors.white)),
-                onTap: () {
-                  // Update the state of the app
-                  // ...
-                  // Then close the drawer
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                        builder: (context) => const AssistidosPage()),
-                  );
-                },
-              ),
-              ListTile(
-                leading: const Icon(Icons.download, color: Colors.white),
-                title: const Text('Relatórios',
-                    style: TextStyle(color: Colors.white)),
-                onTap: () {
-                  // Update the state of the app
-                  // ...
-                  // Then close the drawer
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                        builder: (context) => const ReportsPage()),
-                  );
-                },
-              ),
-              ListTile(
-                leading: const Icon(Icons.calendar_month, color: Colors.white),
-                title:
-                    const Text('Agenda', style: TextStyle(color: Colors.white)),
-                onTap: () {
-                  // Update the state of the app
-                  // ...
-                  // Then close the drawer
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                        builder: (context) => const SchedulePage()),
-                  );
-                },
-              ),
-              ListTile(
-                leading: const Icon(Icons.bar_chart, color: Colors.white),
-                title: const Text('Estatisticas',
-                    style: TextStyle(color: Colors.white)),
-                onTap: () {
-                  // Update the state of the app
-                  // ...
-                  // Then close the drawer
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                        builder: (context) => const StatisticsPage()),
-                  );
-                },
-              ),
-              ListTile(
-                leading: const Icon(Icons.settings, color: Colors.white),
-                title: const Text('Configurações',
-                    style: TextStyle(color: Colors.white)),
-                onTap: () {
-                  // Update the state of the app
-                  // ...
-                  // Then close the drawer
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                        builder: (context) => const SettingsPage()),
-                  );
-                },
-              )
-            ],
-          ),
-        ),
-      ),
+      drawer: const SideMenu(),
     );
   }
 }
@@ -433,10 +302,13 @@ class MyData2 extends DataTableSource {
 
   @override
   bool get isRowCountApproximate => false;
+
   @override
   int get rowCount => _data.length;
+
   @override
   int get selectedRowCount => 0;
+
   @override
   DataRow getRow(int index) {
     return DataRow(cells: [
